refactor(cleanup): split structural hotspots
