+++ tests/services/test_alignment_ffmpeg.py
@@ -1,0 +1,150 @@
+"""FFmpeg and ffprobe audio alignment tests."""
+
+# pyright: reportPrivateUsage=false
+
+from fractions import Fraction
+from pathlib import Path
+from unittest.mock import MagicMock, patch
+
+import pytest
+
+from frame_compare.services.alignment_audio import (
+    extract_audio as _extract_audio,
+)
+from frame_compare.services.alignment_audio import (
+    probe_fps as _probe_fps,
+)
+from frame_compare.services.errors import AudioAlignmentError
+from frame_compare.utils.ffmpeg_errors import FFmpegError, FFmpegNotFoundError
+
+
+@patch("frame_compare.services.alignment_audio.run_subprocess")
+def test_probe_fps_fraction(mock_run: MagicMock):
+    """Test probing FPS when it returns a fraction."""
+    mock_run.return_value.stdout = b"24000/1001\n"
+    res = _probe_fps(Path("test.mkv"))
+    assert res == Fraction(24000, 1001)
+    mock_run.assert_called_once_with(
+        [
+            "ffprobe",
+            "-v",
+            "quiet",
+            "-select_streams",
+            "v:0",
+            "-show_entries",
+            "stream=r_frame_rate",
+            "-of",
+            "csv=p=0",
+            "test.mkv",
+        ],
+        timeout_seconds=15.0,
+    )
+
+
+@patch("frame_compare.services.alignment_audio.run_subprocess")
+def test_probe_fps_integer(mock_run: MagicMock):
+    """Test probing FPS when it returns an integer."""
+    mock_run.return_value.stdout = b"24\n"
+    res = _probe_fps(Path("test.mkv"))
+    assert res == Fraction(24, 1)
+
+
+@patch("frame_compare.services.alignment_audio.run_subprocess")
+def test_probe_fps_empty_output_preserves_ffmpeg_error(mock_run: MagicMock):
+    """Empty ffprobe output should preserve the original FFmpegError details."""
+    proc = MagicMock()
+    proc.stdout = b""
+    proc.returncode = 7
+    mock_run.return_value = proc
+
+    with pytest.raises(FFmpegError) as exc_info:
+        _probe_fps(Path("test.mkv"))
+    assert exc_info.value.context.details is not None
+    assert exc_info.value.context.details.get("returncode") == 7
+    assert "ffprobe returned empty output" in str(exc_info.value.context.details.get("stderr", ""))
+
+
+@patch("frame_compare.services.alignment_audio.run_subprocess")
+def test_probe_fps_not_found_raises(mock_run: MagicMock):
+    """Test probing FPS when ffprobe is missing."""
+    mock_run.side_effect = FileNotFoundError()
+    with pytest.raises(FFmpegNotFoundError):
+        _probe_fps(Path("test.mkv"))
+
+
+@patch("frame_compare.services.alignment_audio.run_subprocess")
+def test_probe_fps_nonzero_exit_raises(mock_run: MagicMock):
+    """Test probing FPS when ffprobe fails."""
+    from subprocess import CalledProcessError
+
+    mock_run.side_effect = CalledProcessError(1, ["ffprobe"], stderr=b"error")
+    with pytest.raises(FFmpegError):
+        _probe_fps(Path("test.mkv"))
+
+
+@patch("frame_compare.services.alignment_audio.run_subprocess")
+def test_extract_audio_ffmpeg_not_found(mock_run: MagicMock):
+    """Test audio extraction when ffmpeg is missing."""
+    mock_run.side_effect = FileNotFoundError()
+    with pytest.raises(FFmpegNotFoundError):
+        _extract_audio(Path("test.mkv"), 8000)
+
+
+@patch("frame_compare.services.alignment_audio.run_subprocess")
+def test_probe_fps_timeout_raises(mock_run: MagicMock):
+    """Test probing FPS timeout surfaces as FFmpegError."""
+    from subprocess import TimeoutExpired
+
+    mock_run.side_effect = TimeoutExpired(cmd=["ffprobe"], timeout=15.0)
+    with pytest.raises(FFmpegError) as exc_info:
+        _probe_fps(Path("test.mkv"))
+    assert exc_info.value.context.details is not None
+    assert exc_info.value.context.details.get("returncode") == 124
+    assert "timed out" in str(exc_info.value.context.details.get("stderr", ""))
+
+
+@patch("frame_compare.services.alignment_audio.run_subprocess")
+def test_extract_audio_ffmpeg_fails(mock_run: MagicMock):
+    """Test audio extraction when ffmpeg fails."""
+    from subprocess import CalledProcessError
+
+    mock_run.side_effect = CalledProcessError(1, ["ffmpeg"], stderr=b"error")
+    with pytest.raises(FFmpegError):
+        _extract_audio(Path("test.mkv"), 8000)
+
+
+@patch("frame_compare.services.alignment_audio.run_subprocess")
+def test_extract_audio_timeout_raises(mock_run: MagicMock):
+    """Test audio extraction timeout surfaces as FFmpegError."""
+    from subprocess import TimeoutExpired
+
+    mock_run.side_effect = TimeoutExpired(cmd=["ffmpeg"], timeout=120.0)
+    with pytest.raises(FFmpegError) as exc_info:
+        _extract_audio(Path("test.mkv"), 8000)
+    assert exc_info.value.context.details is not None
+    assert exc_info.value.context.details.get("returncode") == 124
+    assert "timed out" in str(exc_info.value.context.details.get("stderr", ""))
+
+
+@patch("frame_compare.services.alignment_audio.run_subprocess")
+def test_extract_audio_empty_raises(mock_run: MagicMock):
+    """Test audio extraction when output is empty."""
+    mock_run.return_value.stdout = b""
+    with pytest.raises(AudioAlignmentError, match="empty audio"):
+        _extract_audio(Path("test.mkv"), 8000)
+    mock_run.assert_called_once_with(
+        [
+            "ffmpeg",
+            "-i",
+            "test.mkv",
+            "-vn",
+            "-ac",
+            "1",
+            "-ar",
+            "8000",
+            "-f",
+            "f32le",
+            "-",
+        ],
+        timeout_seconds=120.0,
+    )
